+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Excessive polling/crashes with low ModemManager refresh rate
+- Search not using latest GPS coordinate as reference
 
 ## 1.3.0 - 2026-01-13